Move chart configurations into stories
