+++ stories/charts/box_plot_group.ts
