+++ stories/charts/box_plot_group_numbers.ts
