+++ stories/charts/box_plot_numbers.ts
