+++ stories/charts/floating_bar.ts
