+++ stories/charts/getting_started_Example.ts
