+++ stories/charts/grouped_bar.ts
