+++ stories/charts/grouped_line.ts
