+++ stories/charts/grouped_scatter.ts
