+++ stories/charts/simple_bar.ts
