+++ stories/charts/stacked_bar.ts
