+++ stories/charts/string_x_values.ts
